% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_xref.R
\name{clean_indi_names}
\alias{clean_indi_names}
\title{Temporarily remove forward slashes from surnames}
\usage{
clean_indi_names(gedcom)
}
\arguments{
\item{gedcom}{A tidyged object.}
}
\value{
A tidyged object with all forward slashes removed from surnames.
}
\description{
Temporarily remove forward slashes from surnames
}
