% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_value.R
\name{indi_name}
\alias{indi_name}
\title{Get an individual's name}
\usage{
indi_name(gedcom, xref)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{An xref of an Individual record.}
}
\value{
The name of the individual. If one is not defined, the xref is returned.
}
\description{
Get an individual's name
}
